requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.4.0
django-admin-interface==0.26.0 